0x0C-python-almost_a_circle/models/rectangle.py
#!/usr/bin/python3
from models.base import Base


class Rectangle(Base):
    """
    """

    def __init__(self, width, height, x=0, y=0, id=None):
        """
        Rectangle class constructor.

        Args:
            width (int): Width of the rectangle.
            height (int): Height of the rectangle.
            x (int, optional): Horizontal position. Defaults to 0.
            y (int, optional): Vertical position. Defaults to 0.
            id (int, optional): Identifier. Defaults to None.
        """
        super().__init__(id)
        self.width = width
        self.height = height
        self.x = x
        self.y = y

    @property
    def width(self):
        """Getter for width"""
        return self.__width

    @width.setter
    def width(self, value):
        """Setter for width with validation"""
        if not isinstance(value, int):
            raise TypeError("width must be an integer")
        if value <= 0:
            raise ValueError("width must be greater than 0")
        self.__width = value

    @property
    def height(self):
        """Getter for height"""
        return self.__height

    @height.setter
    def height(self, value):
        """Setter for height with validation"""
        if not isinstance(value, int):
            raise TypeError("Height must be an integer")
        if value <= 0:
            raise ValueError("Height must be greater than 0")
        self.__height = value

    @property
    def x(self):
        """Getter for x."""
        return self.__x

    @x.setter
    def x(self, value):
        """Setter for x with validation."""
        if not isinstance(value, int):
            raise TypeError("X must be an integer")
        self.__x = value

    @property
    def y(self):
        """Getter for y."""
        return self.__y

    @y.setter
    def y(self, value):
        """Setter for y with validation."""
        if not isinstance(value, int):
            raise TypeError("Y must be an integer")
        self.__y = value
